refactor: Move view and stream from `datasource` to `catalog`, deprecate `View::try_new` (#15260)
* First Iteration

* minor change

* unnecessary addition removal

* sorting weird imports

* fix: doc

* fix:doctest

* apply suggestions

* fix:doctest

* Deprecate View::try_new w/ hint

---------
