+++ datafusion/catalog/Cargo.toml
@@ -35,8 +35,10 @@ arrow = { workspace = true }
 async-trait = { workspace = true }
 dashmap = { workspace = true }
 datafusion-common = { workspace = true }
+datafusion-common-runtime = { workspace = true }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
+datafusion-physical-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
 datafusion-sql = { workspace = true }
 futures = { workspace = true }
@@ -44,8 +46,6 @@ itertools = { workspace = true }
 log = { workspace = true }
 object_store = { workspace = true }
 parking_lot = { workspace = true }
-
-[dev-dependencies]
 tokio = { workspace = true }
 
 [lints]
