+++ datafusion/physical-expr/src/lib.rs
@@ -54,8 +54,8 @@ pub use equivalence::{
 };
 pub use partitioning::{Distribution, Partitioning};
 pub use physical_expr::{
-    physical_exprs_bag_equal, physical_exprs_contains, physical_exprs_equal,
-    PhysicalExprRef,
+    create_ordering, physical_exprs_bag_equal, physical_exprs_contains,
+    physical_exprs_equal, PhysicalExprRef,
 };
 
 pub use datafusion_physical_expr_common::physical_expr::PhysicalExpr;
